dxf import: NURBS bugfix
http://www.thingiverse.com/thing:52386 imports with correct shape now

The changes are almost an 1:1 translation from rbasis.c of the examaple
source code cited in the comment, but some array indexes are shifted.

Only one DXF example was used, so it's possible that I added a new
mistake; I just tried to translate the C code 1:1.

In particular, weights are untested.