new raster protocol
- new serial protocol and buffering for raster data
- new "pixels per mm" setting
  - stepper ISR keeps track of (fractional) distance to next pixel
- raster resolution doesn't have to match the microstep resolution
- raster moves can have any direction/angle now

Raster data is no longer streamed via the serial interrupt while the
move is already in progress. Instead it's expected right after the
"line" command if the new "raster bytes" parameter was used.  The
move is then queued by the planner together with its raster buffer.

The raster buffer is limited to 60 bytes, which will take long enough
to execute to allow the CPU to schedule the next move and combine it
without slowing down while cruising at high speed.

The block buffer size was decreased to make room for 4 raster buffers.
I think this will not become a problem, but if it does, we should
consider to enlarge the serial RX buffer instead, because it buffers
commands with fewer bytes per move and it stores raster data too.

New raster_line() command in driveboard.py to split long raster moves
into smaller ones.