stepper.c performance optimization
The stepper ISR usually takes about 25us to complete, but when
adjust_speed() is called it can take 150us instead.

This is a problem because it can prevent the next stepper ISR from
happening, which effectively alters the number of microsteps per
second, which is bad for time-based raster (and for higher speeds).

This commit shaves off 55us from adjust_speed().

Numbers to compare with:

 4'000 mm/min = 165us between microsteps
 8'000 mm/min =  82us between microsteps
12'000 mm/min =  55us between microsteps