firmware: fix occasional missing CHUNK_PROCESSED
When a large amount of data is read by the firmware at once, the
firmware may generate CHUNK_PROCESSED faster than they can be sent,
so we have to count the requests.

This mainly happens when copying raster data into the buffer, and
is more likely to happen if the status is polled very often because
this blocks the firmware from processing raster data as it arrives.